\name{RShiny-package}
\alias{RShiny-package}
\alias{RShiny}
\docType{package}
\title{
\packageTitle{RShiny}
}
\description{
\packageDescription{RShiny}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{RShiny}
\packageIndices{RShiny}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{RShiny}

Maintainer: \packageMaintainer{RShiny}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
